+++ packages/solid/src/components/field/field.test.tsx
@@ -39,6 +39,9 @@ describe('Field / Input', () => {
   it('should set textbox as disabled', async () => {
     render(() => <ComponentUnderTest disabled />)
     expect(screen.getByRole('textbox', { name: /label/i })).toBeDisabled()
+    expect(document.querySelector('[data-part="root"]')).toHaveAttribute('data-disabled')
+    expect(screen.getByText('Label')).toHaveAttribute('data-disabled')
+    expect(screen.getByText('Some additional Info')).toHaveAttribute('data-disabled')
   })
 
   it('should set textbox as readonly', async () => {