+++ packages/vue/src/components/field/tests/field.test.tsx
@@ -26,6 +26,9 @@ describe('Field', () => {
   it('should set textbox as disabled', async () => {
     render(ComponentUnderTest, { props: { disabled: true } })
     expect(screen.getByRole('textbox', { name: /label/i })).toBeDisabled()
+    expect(document.querySelector('[data-part="root"]')).toHaveAttribute('data-disabled')
+    expect(screen.getByText('Label')).toHaveAttribute('data-disabled')
+    expect(screen.getByText('Some additional Info')).toHaveAttribute('data-disabled')
   })
 
   it('should set textbox as readonly', async () => {